requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.9.0
mysql-connector-python==8.2.0
pydantic==2.5.0
sse-starlette==1.8.2
mcp==0.1.0
python-dotenv==1.0.0